feat: hook up verified name status in place of is_verified
Everywhere that used to manipulate the is_verified flag needs to be
able to manipulate status.

Once the verified name has a status is_verified is just status =
approved.

is_verified lives on in the get since the common case for clients is
to want only the most recent approved name and is_verified is an OK
name for that.