feat: Add verified_name_enabled and use_verified_name_for_certs to the GET response of VerifiedNameHistoryView.
The Account Settings page will transition to using the VerifiedNameHistoryVieew instead of the VerifiedNameView. Therefore, information about whether the verified name feature is enabled and whether a user's verified name is selected for use by certificates is relevant information that should be passed in the response.

MST-954 (https://openedx.atlassian.net/browse/MST-954)