fix: upgrade pip-tools to fix bug in versions 6.6.0 and 6.6.1
This commit upgrades the version of pip-tools used in this repository from 6.6.0 to 6.6.2.

In version 6.6.0 of pip-tools, there is a bug that is preventing pip-tools from working. This is breaking the Python requirements update GitHub action in this repository.

The error is "ImportError: cannot import name 'BAR_TYPES' from 'pip._internal.cli.progress_bars'". The error was reported here: jazzband/pip-tools#1617. The fix to this bug was released in version 6.6.2 of pip-tools. See the comment here: jazzband/pip-tools#1617 (comment).

Version 6.6.1 of pip-tools also has a bug, which is fixed in version 6.6.2. I observed this issue breaking the Python requirements update GitHub action in another repository, so I have upgrade the version straight to 6.6.2. The issue in version 6.6.1 was reported here: jazzband/pip-tools#1624.